# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Course Information – API for course memos, used for web-based memos",
    "description": "API that supports the course memo administraion and public site that makes it possible to create and manage course memos.",
    "version": "1.0.0"
  },
  "basePath": "/api/kurs-pm-data",
  "produces": ["application/json"],
  "paths": {
...
    "/v1/pdfCourseMemoFilesList/{courseCode}/{semester}": {
      "get": {
        "operationId": "getStoredMemoPdfListByCourseCode",
        "summary": "Get a list of course memos stored as pdf files (integrated from kurs-pm-api) used by kursinfo-web",
        "description": "Gets course memos for rounds in a course. Memos stored as pdfs in kursinfo storage (not generated)",
        "parameters": [
          {
            "name": "courseCode",
            "in": "path",
            "description": "Id of the course code",
            "required": true,
            "type": "string"
          },
          {
            "name": "semester",
            "in": "path",
            "description": "Semester code, five numbers as a string (year + 1 (spring) or 2 (fall). Filter so the result only has memo data from that semester and all subsequent semesters",
            "required": false,
            "type": "string"
          }
        ],
        "tags": ["v1"],
        "responses": {
          "200": {
            "description": "The requested data",
            "schema": {
              "$ref": "#/definitions/MemoPdfFilesList"
            }
          },
          "400": {
            "description": "Bad Request, something with the request is invalid",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "404": {
            "description": "No organization or user matched the Ids",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "default": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        },
        "security": [
          {
            "api_key": ["read"]
          }
        ]
      }
    },
    "/v1/prioritizedWebOrPdfPublishedMemos/{courseCode}/": {
      "get": {
        "operationId": "getPrioritizedWebOrPdfMemosByCourseCode",
        "summary": "Get and prioritize which document to show for each course round, web-based version is prioritized. If there is two versions for the same round, it will choose a web-based.",
        "description": "For each course round search a document, if there is a web-based version then skip a memo saved as a file. One round - one document.",
        "parameters": [
          {
            "name": "courseCode",
            "in": "path",
            "description": "Id of the course code",
            "required": true,
            "type": "string"
          }
        ],
        "tags": ["v1"],
        "responses": {
          "200": {
            "description": "The requested data",
            "schema": {
              "$ref": "#/definitions/allMixedAndPrioritedMemosForCourse"
            }
          },
          "400": {
            "description": "Bad Request, something with the request is invalid",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "404": {
            "description": "No organization or user matched the Ids",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "default": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        },
        "security": [
          {
            "api_key": ["read"]
          }
        ]
      }
    },
    "/v1/webAndPdfPublishedMemos/{courseCode}/": {
      "get": {
        "operationId": "getPdfAndWebMemosListByCourseCode",
        "summary": "Get a list of course memos stored as pdf files (integrated from kurs-pm-api) AND as web-based versions used by kurs-pm-web 'About memo'",
        "description": "Gets both solutions of course memos for a course. Memos stored as pdfs in kursinfo storage (not generated) and as a html version (kurs-pm-data)",
        "parameters": [
          {
            "name": "courseCode",
            "in": "path",
            "description": "Id of the course code",
            "required": true,
            "type": "string"
          }
        ],
        "tags": ["v1"],
        "responses": {
          "200": {
            "description": "The requested data",
            "schema": {
              "$ref": "#/definitions/allMixedMemosForCourse"
            }
          },
          "400": {
            "description": "Bad Request, something with the request is invalid",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "404": {
            "description": "No organization or user matched the Ids",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "default": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        },
        "security": [
          {
            "api_key": ["read"]
          }
        ]
      }
    },
    "/v1/webAndPdfPublishedMemosBySemester/{semester}/": {
      "get": {
        "operationId": "getPdfAndWebMemosListBySemester",
        "summary": "Get a list of course memos for a chosen semester stored as pdf files (integrated from kurs-pm-api) AND as web-based versions used by kursinfo-admin-web 'About memo' statistic",
        "description": "Gets both solutions of course memos for a chosen semester. Memos stored as pdfs in kursinfo storage (not generated) and as a html version (kurs-pm-data)",
        "parameters": [
          {
            "name": "semester",
            "in": "path",
            "description": "Term/Termin/semester Autumn 2020 -> 20202, Spring 2020 -> 20201 (year)(1-2)",
            "required": true,
            "type": "string"
          }
        ],
        "tags": ["v1"],
        "responses": {
          "200": {
            "description": "The requested data",
            "schema": {
              "$ref": "#/definitions/listOfMixedMemosForSemester"
            }
          },
          "400": {
            "description": "Bad Request, something with the request is invalid",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "404": {
            "description": "No organization or user matched the Ids",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "default": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        },
        "security": [
          {
            "api_key": ["read"]
          }
        ]
      }
    },
...
  },
  "securityDefinitions": {
    "api_key": {
      "type": "apiKey",
      "name": "api_key",
      "in": "header",
      "scopes": {
        "read": "Read access to data",
        "write": "Write access to data"
      }
    }
  },
  "definitions": {
...
    "allMixedMemosForCourse": {
      "properties": {
        "courseCode": {
          "type": "string",
          "description": "Course code for which it is valid"
        },
        "miniMemos": {
          "type": "object",
          "description": "memos listed by semester and in a mini version",
          "items": {
            "type": "array",
            "schema": {
              "$ref": "#/definitions/listOfMixedMemosForSemester"
            }
          }
        }
      }
    },
    "allMixedAndPrioritedMemosForCourse": {
      "properties": {
        "courseCode": {
          "type": "string",
          "description": "Course code for which it is valid"
        },
        "miniMemos": {
          "type": "object",
          "description": "memos listed by semester and by round and in a mini version",
          "items": {
            "type": "object",
            "items": {
              "type": "object",
              "schema": {
                "$ref": "#/definitions/listOfMixedMemosForSemester"
              }
            }
          }
        }
      }
    },
    "listOfMixedMemosForSemester": {
      "type": "array",
      "items": "object",
      "properties": {
        "courseCode": {
          "type": "string",
          "description": "Course code for which it is valid"
        },
        "courseMemoFileName": {
          "type": "string",
          "description": "Name of this course memo pdf file in storage (only for pdf-based), usually a string of course offerings names"
        },
        "ladokRoundIds": {
          "type": "array",
          "items": "string",
          "description": "List of round ids (course offerings), including renamed koppsRoundIds if it is a pdf or ladokRoundsIds if it is web-based, that have been used for a course semester"
        },
        "semester": {
          "type": "string",
          "description": "f.e., if it is an autumn 2020: 20202, if it is a spring 2020: 20201"
        },
        "memoEndPoint": {
          "type": "string",
          "description": "memo id which is unchangeable"
        },
        "memoCommonLangAbbr": {
          "type": "string",
          "description": "Language of this course memo (only for web-based)"
        },
        "memoName": {
          "type": "string",
          "description": "Name of this course memo (only for web-based), usually a string of course offerings names"
        },
        "isPdf": {
          "type": "boolean",
          "description": "Is it a pdf or web-based memo?"
        },
        "version": {
          "type": "integer",
          "description": "Version of this course memo (only for web-based)"
        },
        "lastChangeDate": {
          "type": "string",
          "description": "Date and time and hour for last change of this memo "
        },
        "previousFileList": {
          "type": "array",
          "items": "object",
          "description": "Array for object with previous memo files and its upload date",
          "properties": {
            "courseMemoFileName": {
              "type": "string",
              "description": "File name of a previous course memo"
            },
            "lastChangeDate": {
              "type": "string",
              "description": "Published date of a previous course memo"
            },
            "version": {
              "type": "string",
              "description": "Version number of course memo, the earliest version starts with 1"
            }
          }
        }
      }
    },
...
    "MemoPdfFilesList": {
      "properties": {
        "_id": {
          "type": "string",
          "description": "Unique identifier representing some specific data"
        },
        "courseCode": {
          "type": "string",
          "description": "Course code for which it is valid"
        },
        "pdfMemoUploadDate": {
          "type": "string",
          "description": "When document were published (first)?? time"
        },
        "koppsRoundId": {
          "type": "string",
          "description": "List of round id:s for which memo is valid (kurstillfälle)"
        },
        "courseMemoFileName": {
          "type": "string",
          "description": "Uploaded courseMemo file"
        },
        "changedAfterPublishedDate": {
          "type": "string",
          "description": "Date for changes made after published???"
        }
      }
    },
...
    "Error": {
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "fields": {
          "type": "string"
        }
      }
    },
...
  }
}